@@ -5,17 +5,26 @@
   "main": "index.js",
   "scripts": {
     "start": "ts-node-dev src/index.ts",
-    "test": "echo \"Error: no test specified\" && exit 1",
+    "test": "npx mocha --require ts-node/register --require esm src/**/*.test.ts",
     "seed": "npx sequelize db:seed:all"
   },
   "keywords": [],
   "author": "",
   "license": "ISC",
   "dependencies": {
+    "@types/ansi-colors": "^3.2.2",
     "@types/bcrypt": "^3.0.0",
+    "@types/chai": "^4.2.14",
+    "@types/esm": "^3.2.0",
     "@types/express": "^4.17.9",
+    "@types/faker": "^5.1.5",
+    "@types/mocha": "^8.2.0",
+    "@types/supertest": "^2.0.10",
+    "@types/validator": "^13.1.1",
     "bcrypt": "^5.0.0",
+    "esm": "^3.2.25",
     "express": "^4.17.1",
+    "faker": "^5.1.0",
     "inversify": "^5.0.1",
     "inversify-express-utils": "^6.3.2",
     "node-flywaydb": "^3.0.3",
@@ -23,12 +32,16 @@
     "pg-hstore": "^2.3.3",
     "reflect-metadata": "^0.1.13",
     "sequelize": "^6.3.5",
+    "supertest": "^6.0.1",
     "ts-node-dev": "^1.0.0",
-    "typescript": "^4.0.5",
     "uuidv4": "^6.2.5",
     "winston": "^3.3.3"
   },
   "devDependencies": {
-    "sequelize-cli": "^6.2.0"
+    "chai": "^4.2.0",
+    "mocha": "^8.2.1",
+    "sequelize-cli": "^6.2.0",
+    "ts-node": "^9.1.1",
+    "typescript": "^4.1.3"
   }
 }
