@@ -1,14 +1,14 @@
-import { Person } from "../../../shared/domain/Person";
+import { Person } from "../../Person/domain/Person";
 
 export class ExternalPerson extends Person {
   constructor(public externalPersonId: string, person: Person) {
     super(
-        person.personId,
-        person.fullName,
-        person.lastName,
-        person.surName,
-        person.age,
-        person.identifier
+        person._personId,
+        person._fullName,
+        person._lastName,
+        person._surName,
+        person._age,
+        person._identifier
     );
   }
 
