@@ -1,14 +1,10 @@
 import validator from "validator";
 import StringValueObject from "../../../shared/domain/StringValueObject";
 
-class UserEmail extends StringValueObject {
+class PersonFullName extends StringValueObject {
     constructor(value : string) {
         super(value)
     }
-
-    public validateEmail() : boolean{
-        return validator.isEmail(this._value);
-    }
 }
 
-export default UserEmail;
+export default PersonFullName;
