@@ -1,17 +1,17 @@
 import validator from "validator";
 import StringValueObject from "../../../shared/domain/StringValueObject";
 
-class UserPassword extends StringValueObject {
+class TutorCode extends StringValueObject {
     constructor(value : string) {
         super(value)
     }
 
-    public validatePassword() : boolean {
+    public validateTutorCode() : boolean{
         return validator.isLength(this._value,{
-            min: 5,
-            max: 25
+            min: 8,
+            max: 15
         });
     }
 }
 
-export default UserPassword;
+export default TutorCode;
