@@ -1,8 +1,8 @@
 import { Model, DataTypes, Association } from "sequelize";
 import { uuid } from "uuidv4";
-import { beans, Database } from "../../../../core/beans";
-import DEPENDENCY_TYPES from "../../../../core/beans/ioc-types";
-import PersonModel from "../../../../shared/infraestructure/Persistence/PersonModel";
+import { beans, Database } from "../../../../../core/beans";
+import DEPENDENCY_TYPES from "../../../../../core/beans/ioc-types";
+import PersonModel from "../../../../../shared/infraestructure/Persistence/PersonModel";
 import UserTypeModel from "./UserTypeModel";
 const dbInstance = beans.get<Database>(DEPENDENCY_TYPES.Sequelize);
 const { sequelize } = dbInstance;
