@@ -3,7 +3,7 @@ import { uuid } from "uuidv4";
 import { beans, Database } from "../../../core/beans";
 import DEPENDENCY_TYPES from "../../../core/beans/ioc-types";
 import CourseModel from "../../../modules/Course/Infraestructure/Persistence/CourseModel";
-import StudentModel from "../../../modules/Student/Infraestructure/Persistence/StudentModel";
+import StudentModel from "../../../modules/Student/Infraestructure/Persistence/sequelize/StudentModel";
 const dbInstance = beans.get<Database>(DEPENDENCY_TYPES.Sequelize);
 const { sequelize } = dbInstance;
 
