@@ -4,9 +4,9 @@ import { beans, Database } from "../../../core/beans";
 import DEPENDENCY_TYPES from "../../../core/beans/ioc-types";
 import ExternalPersonModel from "../../../modules/ExternalPerson/Infraestucture/Persistence/ExternalPersonModel";
 import SpecializationModel from "../../../modules/Specialization/Infraestructure/Persistence/SpecializationModel";
-import StudentModel from "../../../modules/Student/Infraestructure/Persistence/StudentModel";
+import StudentModel from "../../../modules/Student/Infraestructure/Persistence/sequelize/StudentModel";
 import TutorModel from "../../../modules/Tutor/Infraestructure/Persistence/TutorModel";
-import  UserModel from "../../../modules/User/Infraestructure/Persistence/UserModel";
+import  UserModel from "../../../modules/User/Infraestructure/Persistence/sequelize/UserModel";
 import IdiomModel from "./IdiomModel";
 const dbInstance = beans.get<Database>(DEPENDENCY_TYPES.Sequelize);
 const { sequelize } = dbInstance;
